Refactor Love2D prototype into modules
